Add BlockBase class
Contains common functionality used by the Player block and the Stackable blocks
Has utility functionality for getting positions of cube faces and updating object rigid body properties
