+++ Assets/BlockBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0edfda87bb714c4da7d95e321d35d15
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
